refactor :: move signup.ts into types/auth
-step.type.ts를 signup.ts로 변경
-앞으로 auth에서 쓰일 type들을 고려해 signup.ts 위치 변경
