+++ src/components/auth/authLayout/AuthStepComponentLayout/ContentLayout.tsx
