+++ src/components/auth/authLayout/AuthStepComponentLayout/StepHeader.tsx
