+++ src/screens/auth/signup/steps/AdminCodeStep.tsx
@@ -1,6 +1,6 @@
 import Typography from "@/components/common/typography/Typography"
 import { useEffect } from "react"
-import ContentLayout from "../layout/ContentLayout";
+import ContentLayout from "@/components/auth/authLayout/AuthStepComponentLayout/ContentLayout";
 import DefaultInput from "@/components/auth/authInput/DefaultInput";
 
 interface AdminCodeStepStepProps {
